Eagle Board file with silkscreen and outer dimension enhancements.  Board now passes DRC checks for most board houses
This version of the .brd file has a significantly cleaned up silkscreen, front & back. It is also a bit wider as the board this was forked from had plated holes extending over the sides.  This is technically a castellated hole and costs more to make.  No changes to the core schematic were made.  If requested I can also post the gerbers from this .brd.